Added Pump/Valve 3-4 Remap compile option
git-svn-id: http://brewtroller.googlecode.com/svn/branches/BrewTroller_1_0@203 786cebc4-e6f9-11dd-9c15-fdd50ba39c1e

+++ Menu.pde
@@ -71,7 +71,7 @@ int getChoice(char choices[][19], int numChoices, int iRow) {
   }
 }
 
-int confirmExit() {
+byte confirmExit() {
   clearLCD();
   printLCD(0, 0, "Exiting will reset");
   printLCD(1, 0, "outputs, setpoints");
